nixos/guix: using systemd features to replace activation script
We're leveraging systemd's tmpfiles to define the directories we want
to create and the paths we want to simlimk.

We then use a pre-start service hook to copy the bootstrap guix store
and set the default binary cache substituer. This script is only
supposed to be run after a guix version change. We store the current
guix store path in a file to keep track of such a change.

For /var/guix, we only symlink the paths we are interested in. This
prevents us to erase any stateful thing (gcroot, profile or sqlite DB)
during a service restart.