power-profiles-daemon: 0.13 -> 0.20
The upstream original maintainer is not paid anymore to maintain the
project and decided to step down. The project has been taken over by
the upower team.

This is the first release part of this upower team.

The integration tests are now relying on the powerprofilectl command:
we need to patch the python path of this dependency before running the
integration tests instead of doing that during in the fixup phase.

Taking advantage of this to remove the PostFixup phase. Not 100% about
this move: it's a bit less future proof.